requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
slackclient==2.9.4
sqlparse==0.4.4
typing_extensions==4.7.0
yarl==1.9.2
